requirements.txt
streamlit
openai
python-dotenv
reportlab
pandas